content/API_Project/requirements.txt
requests==2.31.0
loguru==0.7.3
pytest==7.4.3
allure-pytest==2.13.2
pytest-rerunfailures==13.0
pymysql==1.1.2
DBUtils==3.1.2
Flask==3.1.3
cryptography==46.0.7